Squashed commit of the following:
commit 6ed69caa9d688411d94d0428477ed5cf64ca7f81
Author: Jeromy Cannon <jeromy@swirldslabs.com>
Date:   Thu Feb 6 18:10:36 2025 +0000

    fixed some issues after rebase


commit 9f6d76339f7ba4fcb496712202ad36cbe80ed052
Author: Jeromy Cannon <jeromy@swirldslabs.com>
Date:   Thu Feb 6 18:05:19 2025 +0000

    Squashed commit of the following:

    commit 6b23776ccb6c21753df4a8c16d1dd57e7071013b
    Author: Jeromy Cannon <jeromy@swirldslabs.com>
    Date:   Thu Feb 6 17:59:34 2025 +0000

        implemented k8clientpod and removed some todos


    commit 1826fcf15064e53f3da6900baef457cbc18c7276
    Author: Jeromy Cannon <jeromy@swirldslabs.com>
    Date:   Thu Feb 6 15:59:51 2025 +0000

        renamed K8ClientBase to K8ClientFilter


    commit a961003f868866b8fbeecd86d8bde9e6d75a79ce
    Author: Jeromy Cannon <jeromy@swirldslabs.com>
    Date:   Wed Feb 5 23:52:53 2025 +0000

        first draft


    commit c7fe30b5f2f4d4bfc154ebc0ac9c9bb4f81911e5
    Author: Jeromy Cannon <jeromy@swirldslabs.com>
    Date:   Wed Feb 5 23:52:41 2025 +0000

        first draft
